# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "T-Rader API",
    "description": "日本株投資プラットフォーム T-Rader の API",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/account/risk-policy": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "口座全体のセクター集中度上限 (`max_sector_ratio`) を取得。未設定なら null を返す",
        "operationId": "get_account_risk_policy",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRiskPolicyResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "account"
        ],
        "summary": "口座全体のセクター集中度上限 (`max_sector_ratio`) を更新 (upsert)。`null` で上限を解除する",
        "operationId": "put_account_risk_policy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutAccountRiskPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRiskPolicyResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type ヘッダが application/json ではない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/strategies/{id}/news": {
      "get": {
        "tags": [
          "news"
        ],
        "summary": "戦略に関連付けられたニュース一覧を取得する",
        "description": "`news_strategy_link` を介して `strategy_id` で絞り込み、`published_at` 降順で返す。\n他戦略のニュースは含まれない (link 経由のためテーブル境界で隔離される)。",
        "operationId": "list_strategy_news",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "戦略 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "関連ニュース",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyNewsItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "リクエストパラメータが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountRiskPolicyResponse": {
        "type": "object",
        "properties": {
          "max_sector_ratio": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "API エラーレスポンスの JSON 構造",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "エラーメッセージ"
          }
        }
      },
...
      "MatchedRef": {
        "type": "object",
        "required": [
          "ref_kind",
          "ref_id",
          "matched_term"
        ],
        "properties": {
          "matched_term": {
            "type": "string"
          },
          "ref_id": {
            "type": "string"
          },
          "ref_kind": {
            "type": "string"
          }
        }
      },
...
      "PutAccountRiskPolicyRequest": {
        "type": "object",
        "properties": {
          "max_sector_ratio": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "セクターに属する保有銘柄の時価合計 (口座全体) / 口座全体の保有銘柄時価合計 の上限比率。\n(0, 1] の範囲。`null` で上限を解除する"
          }
        },
        "additionalProperties": false
      },
...
      "StrategyNewsItem": {
        "type": "object",
        "description": "戦略ホームの「関連ニュース」セクション用 1 件",
        "required": [
          "id",
          "source",
          "url",
          "title",
          "published_at",
          "matched_refs"
        ],
        "properties": {
          "body_snippet": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "matched_refs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchedRef"
            },
            "description": "この戦略の interest のうち、このニュースに紐付いたものの一覧 (ref_kind:ref_id)"
          },
          "published_at": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}